requirements.txt
# GenomeAMRAnalyzer Production Dependencies
# =======================================
# Optimized for Python 3.9+ with bioinformatics workflows
# All packages tested for compatibility and performance

# Core Python Scientific Stack
numpy>=1.24.0,<2.0.0
pandas>=1.5.0,<3.0.0
scipy>=1.9.0,<2.0.0

# Bioinformatics Core
biopython>=1.80,<2.0.0

# Configuration and File Handling
pyyaml>=6.0,<7.0.0
toml>=0.10.2
configparser>=5.3.0

# Network and Web Requests
requests>=2.28.0,<3.0.0
urllib3>=1.26.0,<2.0.0

# Data Visualization and Plotting
matplotlib>=3.6.0,<4.0.0
seaborn>=0.12.0,<1.0.0
plotly>=5.0.0,<6.0.0

# Network Analysis and Graph Theory
networkx>=3.0.0,<4.0.0

# Machine Learning and Statistics
scikit-learn>=1.2.0,<2.0.0
statsmodels>=0.14.0,<1.0.0

# Progress Bars and CLI Enhancements
tqdm>=4.64.0,<5.0.0
click>=8.1.0,<9.0.0

# HTML Report Generation
jinja2>=3.1.0,<4.0.0
weasyprint>=59.0,<61.0  # For PDF export
markdown>=3.4.0,<4.0.0

# File Format Support
openpyxl>=3.1.0,<4.0.0  # Excel support
xlsxwriter>=3.1.0,<4.0.0

# System Monitoring and Performance
psutil>=5.9.0,<6.0.0
memory-profiler>=0.61.0,<1.0.0

# Development and Build Tools
setuptools>=65.0.0
wheel>=0.38.0
packaging>=21.0,<24.0

# Optional High-Performance Computing
# Uncomment for maximum performance on supported systems:
# numba>=0.58.0,<1.0.0
# dask>=2023.1.0,<2024.0.0
# joblib>=1.3.0,<2.0.0

# Async support for I/O intensive operations
aiofiles>=23.0.0,<24.0.0

# Progress bars for long operations
tqdm>=4.64.0,<5.0.0
# Priority 2 aligners and utilities
# mappy is the minimap2 python binding; optional and may not build on Windows without toolchain/zlib
# mappy>=2.24
# parasail for SIMD alignments
parasail>=1.3.4
# joblib for caching
joblib>=1.2.0
# pyarrow for potential parquet exports
pyarrow>=11.0.0

# Testing and mocking (dev/test dependencies)
pytest>=7.0.0,<9.0.0
pytest-asyncio>=0.21.0,<0.24.0
aioresponses>=0.7.4,<0.8.0
freezegun>=1.2.0,<2.0.0

# Async HTTP client
aiohttp>=3.9.0,<4.0.0